Fix: minikube fails to start sometimes Some builds of minikube-dind fail to start, showing the message:   Etcd took too long to start in the localkube.err log. The problem is a timing issue in the docker build process. During build, if localkube runs long enough to start etcd, it leaves files in /var/lib/localkube/etcd. If those files are present in the final minikube-dind image, etcd will not start. Once removed, etcd will start successfully.
The fix is to remove any race in the build process by ensuring minikube
(and localkube) are stopped, then remove the etcd data directory.

In addition, we now also remove the initial localkube certs and
kubeconfig generated during build, any files in /tmp, and show the
minikube cache during build so it's available in the build log.